requirements.txt
Python==3.8
Django==2.2
django-cors-headers==3.11.0
djangorestframework==3.13.1